Adds toto_start_record and toto_stop_record (WIP)
In some cases it is not feasible to wrap a single command with
toto-run, e.g. when allowedly editing a bunch of files shortly before
building.
To not disrupt the in-toto supply chain we provide two additional cli
tools that, where one only records materials to a temporary file
and the second records products and finalizes the link metadata
file.